LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-media-dynamic-expected.txt



FAIL Custom Functions: @media responds to changes assert_equals: --actual before resize expected "A" but got "--f()"
